test(agents): split native Bash policy coverage
